requirements.txt
mediapipe==0.10.31
opencv-python
Pillow
numpy
flask
flask-cors
